chore: setting menu component file 통일
